Code refactoring for ClamAVTCPClient
